[datakit] Move the Nemotron recipe out of the registry
Keep sources.py as a catalog and place the pre-staged source and normalization chain alongside other Datakit download recipes.
